Move test filters to helpers file
