+++ src/it/scala/smtlib/it/TestHelpers.scala
@@ -12,6 +12,12 @@ import parser.Parser
 import lexer.Lexer
 import printer.RecursivePrinter
 import interpreters._
+import smtlib.trees.Commands.*
+import smtlib.trees.Terms.Term
+import smtlib.trees.Tree
+import smtlib.trees.TreesOps
+import smtlib.theories.FixedSizeBitVectors.BitVectorSort
+import smtlib.theories.FloatingPoint.FloatingPointSort
 
 /** Provide helper functions to test solver interfaces and files.
   *
@@ -90,6 +96,44 @@ trait TestHelpers {
   def executeBitwuzla(file: File)(f: (String) => Unit): Unit = {
     Seq("bitwuzla", "-v", "0", "--lang", "--smt2", file.getPath) ! ProcessLogger(f, s => ())
   }
+
+  ////////////////////////////////////////////////////////////////
+  // Solver specific helpers
+
+  /**
+  * Is test in the fragment supported by bitwuzla?
+  * 
+  * May not contain ints, reals, arrays, ADTs
+  *
+  */
+  def testIsBitwuzlaCompatible(cmds: Seq[Command], formula: Term): Boolean =
+    def isCompat(term: Term): Boolean =
+      def violatingTerm(term: Tree, children: Seq[Boolean]): Boolean = {
+        import smtlib.theories.*
+        term match
+          case FixedSizeBitVectors.BitVectorConstant(_, _) => false // internally contains IntNumeral, but in a valid place
+          case Ints.NumeralLit(_) => true
+          case Reals.DecimalLit(_) => true
+          case FixedSizeBitVectors.Int2BV(_, _) => true
+          case FixedSizeBitVectors.BV2Nat(_) => true
+          // this is enough to rule out the current problematic tests, 
+          // but may need refinement
+          case _ => children.contains(true)
+      }
+      !TreesOps.fold(violatingTerm)(term)
+    
+    def compatibleCommand(cmd: Command): Boolean =
+      cmd match
+        case DeclareFun(_, paramSorts, sort) => (sort +: paramSorts).forall {
+            case BitVectorSort(_) => true
+            case FloatingPointSort(_, _) => true
+            case _ => false
+        }
+        case DeclareSort(_, _) => false
+        case DeclareDatatypes(_) => false
+        case Assert(term) => isCompat(term)
+        case _ => true // may need refinement
+    cmds.forall(compatibleCommand) && (isCompat(formula))
 
 }
